Refactor subagent runtime types and targets
